requirements.txt
pycairo
pillow
svgpathtools
shapely
numpy